refactor(plugins): split official privacy filter runtime
